+++ packages/manager/src/components/EntityIcon/EntityIcon.stories.mdx
@@ -1,7 +1,7 @@
 import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import EntityIcon, { EntityIconStory } from './EntityIcon';
+import EntityIconStory from './EntityIconStory';
 
-<Meta title="Core/Icons" component={EntityIcon} />
+<Meta title="Core/Icons" component={EntityIconStory} />
 
 # Entity Icons
